pyproject.toml
[project]
name = "pinformation-bot"
version = "0.0.1"
description = "A simple Discord bot for pinning messages to the bottom of a channel"
authors = [{ name = "dBozee" }]
requires-python = ">=3.14"
license = "MIT"
dependencies = [
    "discord-py~=2.6.4",
    "python-dotenv",
]

[tool.hatch.build.targets.sdist]
include = ["pinformation_bot"]

[tool.hatch.build.targets.wheel]
include = ["pinformation_bot"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
ignore-init-module-imports = true
line-length = 120

[tool.ruff.format]
preview = true
quote-style = "preserve"

[tool.ruff.lint]
select = [
    "B",
    "DTZ",
    "E",
    "F",
    "I",
    "PTH",
    "S",
    "SIM",
    "TID251",
    "UP",
]